[Starts #80559928] Fixed issues with DB being closed, removed close() method on DataAccessManager: it was close to useless, but was causing trouble

+++ test/com/alertavert/sentinel/controllers/ApiControllerSpec.scala
@@ -36,9 +36,11 @@ class ApiControllerSpec extends PlaySpec with Results with OneAppPerSuite with B
   }
 
   before {
-    MongoUserDao().clear()
-    MongoOrganizationDao().clear()
-    UserOrgsAssocDao().collection.drop()
+    if (DataAccessManager.isReady) {
+      MongoUserDao().clear()
+      MongoOrganizationDao().clear()
+      UserOrgsAssocDao().collection.drop()
+    }
   }
 
   /**
